64 bit support found here: http://tiger.cs.tsinghua.edu.cn/Students/yangjl/memtime/
On their website, Yang and Li state:
"
memtime is a friendly tool for measuring memory usage which is fetched from a PID information of /proc/[PID]/stat by sampling. It is originally written by Johan Bengtsson in 2002.
It can be found in http://www.update.uu.se/~johanb/memtime/ or http://freecode.com/projects/memtime
But there are some limits or bugs when running in 32 bit linux platforms. Because there are overflows
   - does not allow/report more than 4GB memory usage
   - does not allow runs of longer than 21666 seconds
We have found that unsigned int is used as a counter among the programs which is a upper limit for memory and runtime measuring when run on 64 bit linux platforms. So we fix this by using long int as a counter which can allow us to measure more than 4GB memory usage and 21666 seconds.
Contact

Jianlei Yang, EDA Lab of Tsinghua University, jerryyangs AT gmail.com
Zuowei Li, EDA Lab of Tsinghua University, lizuoweirain AT gmail.com

June 1st, 2012
Tsinghua Design Automation
"

I just collected their patch, credits go to Zuowei Li and Jianlei Yang

+++ machdep.h
@@ -24,12 +24,24 @@
 
 #ifndef MACHDEP_H
 #define MACHDEP_H
+
+#include <sys/time.h>
+#include <asm/param.h>
+#include <sys/stat.h>
+#include <unistd.h>
+#include <fcntl.h>
+
+#include <stdio.h>
+#include <string.h>
+
+#include <sys/resource.h>
+
  
 struct memtime_info {
      unsigned int utime_ms;
      unsigned int stime_ms;
-     unsigned int rss_kb;
-     unsigned int vsize_kb;
+     long int rss_kb;
+     long int vsize_kb;
 };
 
 int init_machdep(pid_t process);
